pyproject.toml
[build-system]
requires = [
    "setuptools >=65.0",
    "wheel >=0.36.2",
]
build-backend = "setuptools.build_meta"

[project]
name = "app"
dynamic = ["version"]
readme = "README.md"
license = {text = 'See LICENSE'}
description = "app"
requires-python = ">=3.10"
dependencies = [
    "modal-client >=0.0.1",
]
[[project.authors]]
name = "Anthony Corletti"
email = "anthcor@gmail.com"

[project.optional-dependencies]
test = [
    "pytest >=6.2.5",
    "coverage >=6.1.1",
    "pytest-cov >=3.0.0",
]
dev = [
    "mypy >=0.910",
    "black >=21.10b0",
    "ruff >=0.0.98",
    "pre-commit >=2.17.0",
]

[project.urls]
Documentation = "https://github.com/anthonycorletti/modal-arbitrary-code-example"

[tool.setuptools.dynamic]
version = {attr = "app.__version__"}

[tool.ruff]
select = ["E", "F", "I"]
ignore = ["D10"]
line-length = 88
target-version = "py310"
extend-ignore = [
    "D203",
    "D204",
    "D213",
    "D215",
    "D400",
    "D404",
    "D406",
    "D407",
    "D408",
    "D409",
    "D413",
    "D415",
]
